+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/bit.rb
@@ -2,7 +2,11 @@ module ActiveRecord
   module ConnectionAdapters
     module PostgreSQL
       module OID # :nodoc:
-        class Bit < Type::String
+        class Bit < Type::Value
+          def type
+            :bit
+          end
+
           def type_cast(value)
             if ::String === value
               case value