+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -104,7 +104,9 @@ module ActiveRecord
         json:        { name: "json" },
         ltree:       { name: "ltree" },
         citext:      { name: "citext" },
-        point:       { name: "point"}
+        point:       { name: "point"},
+        bit:         { name: "bit"},
+        bit_varying: { name: "bit varying" }
       }
 
       OID = PostgreSQL::OID #:nodoc:
@@ -434,7 +436,7 @@ module ActiveRecord
           m.alias_type 'bpchar', 'varchar'
           m.register_type 'bool', Type::Boolean.new
           m.register_type 'bit', OID::Bit.new
-          m.alias_type 'varbit', 'bit'
+          m.register_type 'varbit', OID::BitVarying.new
           m.alias_type 'timestamptz', 'timestamp'
           m.register_type 'date', OID::Date.new
           m.register_type 'time', OID::Time.new